+++ ChangeLog.md
@@ -3,8 +3,9 @@ OpenAI APIs for XP ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Added tracing capabilities to `com.openai.rest.OpenAIEndpoint` class
+  (@thekid)
 * Merged PR #2: Implement REST API - including support for streaming
   (@thekid)
 * Merged PR #1: Add integration tests for `cl100k_base` and `o200k_base`
   (@thekid)
-* First public release - (@thekid)